chore: remove generator-react-component
`generator-react-component` is a terrific tool. But it's too much overhead for this project. In particular, it was painful to publish the site with the default publishing task. I had to keep adding a CNAME file. Additionally, people have managed to find it on github and start contributing. I'd like to make that as friction-free as possible.

Before this commit, all code was generated using `generator-react-component` bundled tasks.

This is because it was the easiest way for me to stop talking about it and start building it.

Now, all files will live as shollowly as possible, built with webpack.