style: restyle using new css libs
This commit is 2-fold.
It paves the way for this lib to be used as an honest-to-goodnes npm
module.

The `ReactCheatSheet` component now uses render callbacks to delegate
styling and rendering of the data.

In an effort to demonstrate this, I changed some of the styles.

The visible changes are as follows:

* used btn.css to style the buttons
* used open-color.css for all the colors
* used point.css for layout
* used minions.css for all other css properties